Changes made while writing the build instructions.

+++ db/migrate/20131109194149_drop_animal_food.rb
@@ -1,6 +1,6 @@
 class DropAnimalFood < ActiveRecord::Migration
   def change
-    drop_table :animal_foods
+    drop_table :animal_foods if self.table_exists?("animal_foods")
     
     add_column :animals, :food_id, :integer
   end
